Adding padding
Add padding to operating system logic (by filling the file with zeros) to 4608 bytes (512*9). This is done to make kernel.img run on bochs and virtual box. Also, it is not necessary to adjust the number of sectors to read from floppy disk until the operating system logic is greater than 4608 bytes.